+++ src/Service/CloudFormation/CHANGELOG.md
@@ -7,6 +7,7 @@
 - AWS api-change: Added `us-iso-west-1` region
 - AWS api-change: Use specific configuration for `us` regions
 - AWS enhancement: Documentation updates.
+- Added `describeStackDriftDetectionStatus` method.
 
 ## 1.1.0